+++ src/stronghold/secret.rs
@@ -239,7 +239,7 @@ mod tests {
         );
         let mut stronghold_adapter = StrongholdAdapter::builder()
             .password("drowssap")
-            .try_build(stronghold_path.clone())
+            .build(stronghold_path.clone())
             .unwrap();
 
         stronghold_adapter.store_mnemonic(mnemonic).await.unwrap();
@@ -275,7 +275,7 @@ mod tests {
         );
         let mut stronghold_adapter = StrongholdAdapter::builder()
             .password("drowssap")
-            .try_build(stronghold_path.clone())
+            .build(stronghold_path.clone())
             .unwrap();
 
         stronghold_adapter.store_mnemonic(mnemonic).await.unwrap();
@@ -286,18 +286,16 @@ mod tests {
         stronghold_adapter.clear_key().await;
 
         // Address generation returns an error when the key is cleared.
-        assert!(
-            stronghold_adapter
-                .generate_addresses(
-                    IOTA_COIN_TYPE,
-                    0,
-                    0..1,
-                    false,
-                    GenerateAddressMetadata { syncing: false },
-                )
-                .await
-                .is_err()
-        );
+        assert!(stronghold_adapter
+            .generate_addresses(
+                IOTA_COIN_TYPE,
+                0,
+                0..1,
+                false,
+                GenerateAddressMetadata { syncing: false },
+            )
+            .await
+            .is_err());
 
         stronghold_adapter.set_password("drowssap").await.unwrap();